+++ R/downloader.R
@@ -46,29 +46,4 @@ unset_cgal <- function(...) {
   Sys.unsetenv("CGAL_DIR")
 }
 
-#' CGAL header file installer
-#' 
-#' @description
-#' `r lifecycle::badge("deprecated")`
-#' This function previously installed or updated the CGAL header files. It has been hard deprecated instead of soft deprecated due to CRAN request. 
-#' 
-#'
-#' @param cgal_path Path to CGAL files on your machine, a URL. If NULL will downloaded latest version from GitHub. Default is NULL.
-#' @param version Desired version to search for from the GitHub. If NULL, will download latest.
-#' @param clean_files Whether to remove calls to C std::err from header files to avoid errors in R. Default is TRUE
-#' @param force Whether to force downloading and install the header files if they're already found in the package. Default is FALSE
-#'
-#' @return NULL
-#' 
-#' @export
-#' 
-cgal_install <- function(cgal_path = NULL, version = NULL, clean_files = TRUE, force = FALSE) {
-  
-  #deprecation error
-  lifecycle::deprecate_stop(
-    when = "5.6.0",
-    what = "cgal_install()",
-    details = "Please use the function `set_cgal()` instead to install your desired version of the CGAL  header files. `cgal_install()` is no longer supported due to CRAN policy."
-  )
-}